+++ src/api.rs
@@ -3,6 +3,7 @@ mod color_light_handler;
 mod color_light_strip_handler;
 mod energy_monitoring_plug_handler;
 mod generic_device_handler;
+mod hub_handler;
 mod light_handler;
 mod plug_handler;
 
@@ -11,5 +12,6 @@ pub use color_light_handler::*;
 pub use color_light_strip_handler::*;
 pub use energy_monitoring_plug_handler::*;
 pub use generic_device_handler::*;
+pub use hub_handler::*;
 pub use light_handler::*;
 pub use plug_handler::*;