+++ src/requests.rs
@@ -2,10 +2,7 @@
 
 mod color;
 mod energy_data_interval;
-mod get_device_info;
-mod get_device_usage;
 mod get_energy_data;
-mod get_energy_usage;
 mod handshake;
 mod lighting_effect;
 mod login_device;
@@ -18,10 +15,7 @@ pub use energy_data_interval::*;
 pub use lighting_effect::*;
 pub use set_device_info::*;
 
-pub(crate) use get_device_info::*;
-pub(crate) use get_device_usage::*;
 pub(crate) use get_energy_data::*;
-pub(crate) use get_energy_usage::*;
 pub(crate) use handshake::*;
 pub(crate) use login_device::*;
 pub(crate) use secure_passthrough::*;