+++ src/responses.rs
@@ -1,5 +1,7 @@
 //! Tapo response objects.
 
+mod child_device_list_result;
+mod decodable_result_ext;
 mod device_info_result;
 mod device_usage_result;
 mod energy_data_result;
@@ -9,10 +11,13 @@ mod tapo_response;
 mod tapo_result;
 mod token_result;
 
+pub use child_device_list_result::*;
 pub use device_info_result::*;
 pub use device_usage_result::*;
 pub use energy_data_result::*;
 pub use energy_usage_result::*;
+
+pub(crate) use decodable_result_ext::*;
 pub(crate) use handshake_result::*;
 pub(crate) use tapo_response::*;
 pub(crate) use tapo_result::*;