+++ src/responses/decodable_result_ext.rs
@@ -3,12 +3,18 @@ use base64::{engine::general_purpose, Engine as _};
 use crate::error::Error;
 
 /// Implemented by all Device Info Result variations.
-pub trait DeviceInfoResultExt
+pub(crate) trait DecodableResultExt
 where
     Self: Sized,
 {
     /// Decodes a base64 encoded string from the result.
-    fn decode(&self) -> Result<Self, Error>;
+    fn decode(self) -> Result<Self, Error>;
+}
+
+impl DecodableResultExt for serde_json::Value {
+    fn decode(self) -> Result<Self, Error> {
+        Ok(self)
+    }
 }
 
 pub(crate) fn decode_value(value: &str) -> anyhow::Result<String> {