+++ src/responses/device_info_result.rs
@@ -1,14 +1,14 @@
 mod default_state;
-mod device_info_result_ext;
 mod generic_device;
+mod hub;
 mod l510;
 mod l530;
 mod l930;
 mod plug;
 
 pub use default_state::*;
-pub use device_info_result_ext::*;
 pub use generic_device::*;
+pub use hub::*;
 pub use l510::*;
 pub use l530::*;
 pub use l930::*;