+++ src/responses/device_info_result/generic_device.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
-use crate::responses::{decode_value, DeviceInfoResultExt, TapoResponseExt};
+use crate::responses::{decode_value, DecodableResultExt, TapoResponseExt};
 
 /// Device info of [`crate::ApiClient<GenericDevice>`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -37,12 +37,11 @@ pub struct GenericDeviceInfoResult {
 
 impl TapoResponseExt for GenericDeviceInfoResult {}
 
-impl DeviceInfoResultExt for GenericDeviceInfoResult {
-    fn decode(&self) -> Result<Self, Error> {
-        Ok(Self {
-            ssid: decode_value(&self.ssid)?,
-            nickname: decode_value(&self.nickname)?,
-            ..self.clone()
-        })
+impl DecodableResultExt for GenericDeviceInfoResult {
+    fn decode(mut self) -> Result<Self, Error> {
+        self.ssid = decode_value(&self.ssid)?;
+        self.nickname = decode_value(&self.nickname)?;
+
+        Ok(self)
     }
 }